+++ lib/api.dart
@@ -1,6 +1,0 @@
-library;
-
-export 'src/endpoint.dart';
-export 'src/error_response.dart';
-export 'src/api_failure.dart';
-export 'src/response.dart';