+++ lib/apiarist.dart
@@ -1,0 +1,7 @@
+library;
+
+export 'src/api_error.dart';
+export 'src/api_failure.dart';
+export 'src/api_endpoint.dart';
+export 'src/api_data_response.dart';
+export 'src/api_response.dart';