+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 1.0.3
+
+- Renamed to class names
+
+## 1.0.2
+
+- Fixed exports, include `ApiError`, remove `ErrorResponse`
+
+## 1.0.1
+
+- Downgraded min Dart SDK (to 3.0) and http (to 1.0.0)
+
 ## 1.0.0
 
 - Initial version.